Copy logic for computing quotient poly
After linearization, the vanishing module is no
longer needed. The prover only needs the logic to
compute the quotient poly.

This commit:
* Deletes the relevant logic from vanishing/prover.rs
* Copies it to prover.rs.
* The copy is 1-to-1, except for a small change in
the return type.
